Minor changes from recent feedback. Needs the 'return repr(str(value))' in '_encode_value' in 'expr_proxy.py', otherwise the 'MyObject()' data for the button 'test_text' does not work.
